LayoutTests/imported/w3c/web-platform-tests/css/css-anchor-position/anchor-name-inline-001-expected.txt

12 123

FAIL .target 1 assert_equals:
<span class="target" data-expected-width="20"></span>
width expected 20 but got 0
FAIL .target 2 assert_equals:
<span class="target" data-expected-width="20"></span>
width expected 20 but got 0
FAIL .target 3 assert_equals:
<span class="target" data-expected-width="30"></span>
width expected 30 but got 0
FAIL .target 4 assert_equals:
<span class="target" data-expected-width="30"></span>
width expected 30 but got 0
FAIL .target 5 assert_equals:
<span class="target" data-expected-width="30"></span>
width expected 30 but got 0
